feat: add support for prompt file in commit command
This commit adds support for specifying a prompt file in the commit command. If a prompt file is provided, the content of the file will be used as the prompt for the chat completion. This allows users to provide a custom prompt for generating chat completions.

The `prompt_file` field has been added to the `Cmd` struct in the `commit` module. If the `prompt_file` option is provided, the content of the file will be read and used as the prompt. If the `prompt_file` option is not provided, the default prompt from the `res/prompt.md` file will be used.

This change enhances the flexibility of the commit command by allowing users to easily customize the prompt used for generating chat completions.